requirements.txt
streamlit==1.41.0
supabase
streamlit-authenticator
python-dotenv
pandas
bcrypt
pytz
euroleague_api
plotly==5.24.1
